+++ test/pond_test.exs
@@ -1,8 +1,5 @@
 defmodule PondTest do
   use ExUnit.Case
   use Pond
-
-  test "greets the world" do
-    assert Pond.hello() == :world
-  end
+  doctest Pond
 end