+++ CHANGELOG.md
@@ -1,6 +1,7 @@
+# ChangeLog
 
 ## v0.1.0
 
 ##### Enhancements
 
-  * Project is born
+  * Add usage examples to README